news-extract/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
filelock==3.12.0
Flask==2.2.3
frozendict==2.3.7
fsspec==2023.4.0
html5lib==1.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
multitasking==0.0.11
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pycparser==2.21
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
yfinance==0.2.18
zipp==3.15.0
